requirements.txt
setuptools==36.5.0
click==6.7
lxml==4.0.0
requests==2.18.4
pandas==0.20.3
pylint==1.7.2
mypy==0.521
mysql-connector==2.2.3
pytest==3.2.2
